learning_backend/pyproject.toml
[project]
name = "learning-backend"
version = "0.1.0"
description = "Minimal Chess960 research backend with rules, engines, environments, and experiment tools."
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "python-chess>=1.999",
  "matplotlib>=3.8",
  "tensorboard>=2.16",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
]

[tool.pytest.ini_options]
testpaths = ["learning_backend/tests"]
pythonpath = ["."]
